Added ParseTree class to handle logging the expansions
ParseTree class can handle logging all of the expansions that are made
for the parser. Parse tree reads in a text file of all the rules. Parse
tree will output the parsing to a local file called parse_tree.txt
inside of the project folder.

Note: each rule in the CFG_Rules.txt file must be on its own line and
that line must be the rule number of that rule for the parse tree to
work properly.